requirements.txt
alembic==1.14.1
iniconfig==2.0.0
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5-stubs==5.15.6.0
PyQt5_sip==12.17.0
pytest==8.3.5
SQLAlchemy==2.0.38
typing_extensions==4.12.2
mistralai~=1.5.1
python-dotenv~=1.0.1
Markdown~=3.7